Crate union_find

Struct and methods for union-find operation.

Structs

QuickFindUf
Union-Find implementation with quick find operation.

QuickUnionUf
Union-Find implementation with quick union operation.

UnionByRank
Operates the union with using the rank of the sets as weight.

UnionByRankSize
Operates the union with using the ramk and the size of the sets as weight.

UnionBySize
Operates the union with using the size of the sets as weight.

UnionBySizeRank
Operates the union with using the size and the rank of the sets as weight.

Enums

UnionResult
Return value of the Union::merge.

Traits

Union
The value that can be contained with Union.

UnionFind
APIs for Union-Find operation.