Struct union_find::UnionFind

pub struct UnionFind { /* fields omitted */ }

Methods

impl UnionFind

fn new(n: usize) -> UnionFind

Create a new partition table with n disjoint partitions numbered 0..n. Running time O(n).

fn union(&mut self, i: usize, j: usize)

Merge the partitions containing i and j. This operation is structured such that the canonical element of the merged partition will be the canonical element of i in the old partition. Running time O(1).

fn find(&mut self, i: usize) -> usize

Return a "canonical element" for the partition containing i. Amortized worst-case running time O(alpha(n)), which is more-or-less O(1). This running time is achieved by mutating the data structure to make future find() operations faster.

fn find_only(&self, i: usize) -> usize

Return a "canonical element" for the partition containing i. Worst-case running time O(n), found in cases involving a sequence of union() and find_only() operations with few find() operations. Prefer find() in any case where it is possible to make the table mutable.

fn same(&mut self, i: usize, j: usize) -> bool

Return true iff i and j are in the same partition. Time complexity is the same as find().

fn same_only(&mut self, i: usize, j: usize) -> bool

Return true iff i and j are in the same partition. Time complexity is the same as find_only(). Prefer same() in any case where it is possible to make the table mutable.