Struct union_find::UnionFind

pub struct UnionFind<'a, T: 'a> {
    pub value: T,
    pub parent: Option<&'a UnionFind<'a, T>>,
}

Also known as the Disjoint-Set data structure.

Some information on UnionFind on Wikipedia.

Creating a set:

 use union_find::UnionFind;
// Create a `UnionFind` node. In order to be useful, a node must be mutable.
let mut x = UnionFind::make_set(1u);
let (mut y, mut z) = (UnionFind::make_set(1u), UnionFind::make_set(2u));
// Exploring behaivor.
assert!(x.value == 1u);
assert!(x.value == y.value);
assert!(x == y); // Gotcha! Use pointers if you need uniques.
assert!(x.value != z.value);
assert!(x != z);
assert!(x.parent == None);

Union two sets:

 use union_find::UnionFind;
 // Some nodes.
 let (mut x, mut y, mut z) = (
  UnionFind::make_set("Foo"),
  UnionFind::make_set("Bar"),
  UnionFind::make_set("Baz"));
 // They are both canonical.
 assert!(x.parent == None);
 assert!(y.parent == None);
 // Union `y` with `x`.
 x.union(&mut y);

Finding the canonical node of a set:

 use union_find::UnionFind;
 let (mut x, mut y, mut z) = (
  UnionFind::make_set("Foo"),
  UnionFind::make_set("Bar"),
  UnionFind::make_set("Baz"));
x.union(&mut y);
// Check relationships.
assert!(*y.find() == x);
assert!(*y.find() == *x.find());
assert!(*y.find() != *z.find());

Fields

value

The value of the node.

parent

Some(parent) for a leaf, or None for a canonical node.

Methods

impl<'a, T> UnionFind<'a, T>

The UnionFind data structure. This is a node that contains a value, and a reference to it's parent, if it has one. All UnionFind nodes must be of a homogenious type.

fn make_set(value: T) -> UnionFind<'a, T>

Encapsulates a value into a UnionFind node. It's parent is set to None, meaning it's a canonical node.

fn find<'a>(&self) -> &'a UnionFind<T>

Fetch the canonical element of the UnionFind dataset containing this one.

fn union<'a>(&'a self, other: &mut UnionFind<'a, T>) -> &'a UnionFind<'a, T>

Union two UnionFind data structures together.